Rewrite emulate button handling
Button handling was heavily changed to try and speed up the emulator but there is still a noticable ~40ms delay that is not present on the hardware.

Also inputs get occasionally dropped but I spent too much time on this to fix it. It's like 90% good now and it'll take 10 hours to get that last 10% so it's not a priority.